Updates
index.html changes complements commit f1aca98 which removes Instagram link replacing it with a mailto link instead.
README.md changes complements commit 70447a8 which was made long time ago, and now includes credit to @makzk due the addition of the OwO blob taken from his page but without adding the orifinal image to the repo itself.
The blob is hosted on owo.cl and not on this repo btw